@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "babel-cli": "^6.24.1",
+    "babel-cli": "^6.26.0",
     "babel-eslint": "^7.2.3",
     "babel-jest": "^20.0.3",
     "babel-plugin-add-module-exports": "^0.2.1",
@@ -13,19 +13,20 @@
     "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.24.1",
-    "babel-runtime": "^6.23.0",
+    "babel-register": "^6.26.0",
+    "babel-runtime": "^6.26.0",
+    "core-js": "^2.5.0",
     "create-react-class": "^15.6.0",
     "cross-env": "^5.0.5",
-    "eslint": "^4.2.0",
+    "eslint": "^4.5.0",
     "eslint-config-google": "^0.9.1",
     "eslint-config-prettier": "^2.3.0",
     "eslint-plugin-flow-vars": "^0.5.0",
     "eslint-plugin-flowtype": "^2.35.0",
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-prettier": "^2.1.2",
-    "eslint-plugin-react": "^7.1.0",
+    "eslint-plugin-prettier": "^2.2.0",
+    "eslint-plugin-react": "^7.2.1",
     "flow-bin": "^0.42.0",
     "glob": "^7.1.1",
     "iflow-debug": "^1.0.15",
@@ -36,7 +37,7 @@
     "lerna": "^2.0.0",
     "plop": "^1.7.4",
     "prettier": "^1.5.3",
-    "prettier-eslint-cli": "^4.1.1",
+    "prettier-eslint-cli": "^4.2.0",
     "prop-types": "^15.5.8",
     "purdy": "^2.2.1",
     "remotedev-server": "^0.2.3",
