@@ -4,10 +4,10 @@ import SiteNav from "../SiteNav"
 import SiteLinks from "../SiteLinks"
 
 class SiteSidebar extends React.Component {
-    render() {
-        const isHome = this.props.location.pathname === ('/');
-        const siteMetadata = this.props.siteMetadata;
-        // TODO, deal with image more nice like
+  render() {
+    const isHome = this.props.location.pathname === `/`
+    const siteMetadata = this.props.siteMetadata
+    // TODO, deal with image more nice like
 
     let header = (
       <div className="">
