@@ -1,7 +1,7 @@
-import React from 'react';
-import * as PropTypes from 'prop-types';
-import Helmet from 'react-helmet';
-import '../static/css/base.scss';
+import React from "react"
+import * as PropTypes from "prop-types"
+import Helmet from "react-helmet"
+import "../static/css/base.scss"
 
 import InsetPage from "./inset-page"
 import BlogPost from "./blog-post"
@@ -12,20 +12,20 @@ class MasterLayout extends React.Component {
     let location = this.props.location.pathname
     let jimmyPage // you jimmy a lock until it opens, so same thing here ;)
 
-      // let dataSource = this.props.pageResources.json.data
-      // let nodeType = dataSource.jsFrontmatter || dataSource.markdownRemark
-      // let frontmatter = nodeType.data || nodeType.frontmatter
-      let passdown = {
-        // frontmatter: frontmatter,
-        location: this.props.location,
-        siteMetadata: siteMetadata,
-        children: this.props.children
-      }
-      if (location === `/` || location === `/contact`) {
-        jimmyPage = <InsetPage {...passdown} />
-      } else {
-        jimmyPage = <BlogPost {...passdown} />
-      }
+    // let dataSource = this.props.pageResources.json.data
+    // let nodeType = dataSource.jsFrontmatter || dataSource.markdownRemark
+    // let frontmatter = nodeType.data || nodeType.frontmatter
+    let passdown = {
+      // frontmatter: frontmatter,
+      location: this.props.location,
+      siteMetadata: siteMetadata,
+      children: this.props.children,
+    }
+    if (location === `/` || location === `/contact`) {
+      jimmyPage = <InsetPage {...passdown} />
+    } else {
+      jimmyPage = <BlogPost {...passdown} />
+    }
 
     return (
       <div className="MasterLayout">
@@ -43,7 +43,6 @@ class MasterLayout extends React.Component {
 }
 
 export default MasterLayout
-
 
 export const pageQuery = graphql`
   query LayoutBySlug {
