@@ -7,9 +7,11 @@ module.exports = {
       resolve: `gatsby-source-mongodb`,
       options: { dbName: `cloud`, collection: `documents`,
                  server: { address: `ds143532.mlab.com`, port: 43532 },
-                 auth: { user: `admin`, password: `12345` }
+                 auth: { user: `admin`, password: `12345` },
+                 map: {description: `text/x-markdown`}
       },
     },
-    `gatsby-plugin-react-helmet`
+    `gatsby-plugin-react-helmet`,
+    `gatsby-transformer-remark`
   ],
 }
