@@ -6,6 +6,7 @@ const _ = require(`lodash`)
 
 const { emitter } = require(`../../redux`)
 const { boundActionCreators } = require(`../../redux/actions`)
+const { getNode } = require(`../../redux`)
 
 function transformPackageJson(json) {
   const transformDeps = deps =>
@@ -69,7 +70,6 @@ exports.sourceNodes = ({ boundActionCreators, store }) => {
       packageJson: transformPackageJson(
         require(`${plugin.resolve}/package.json`)
       ),
-      id: `Plugin ${plugin.name}`,
       parent: `SOURCE`,
       children: [],
       internal: {
@@ -150,5 +150,7 @@ exports.onCreatePage = ({ page, boundActionCreators }) => {
 
 // Listen for DELETE_PAGE and delete page nodes.
 emitter.on(`DELETE_PAGE`, action => {
-  boundActionCreators.deleteNode(createPageId(action.payload.path))
+  const nodeId = createPageId(action.payload.path)
+  const node = getNode(nodeId)
+  boundActionCreators.deleteNode(nodeId, node)
 })
